scripts/数据加工/requirements.txt
open3d
numpy
pillow
opencv-python